feat: :construction: Changed to typescript
Created the skeleton for my typescript application and imported my javascript code. Have not typed code yet.
