--- README.md
@@ -1,1 +1,0 @@
-# DOM-manipulation-and-maps-part-2
