+++ index.html
@@ -22,7 +22,7 @@
         <div id="list"></div>
       </div>
     </main>
-    <script type="module" src="./main.js"></script>
-    <script type="module" src="lap-button.js"></script>
+    <script type="module" src="./main."></script>
+    <script type="module" src="lap-button.ts"></script>
   </body>
 </html>
